Expect and operate in Utf8
Char 160 is (apparently) non-breaking space in win-1252, so operating in
ByteString finds it (C8.isSpace) and trims it. But this results in an
invalid character, visible as "Â" in some contexts.

We can really only do this in Uf8, so let's just expect that. We already
have exception handling (i.e. --strict) for our CRLF check, so any
exceptions trying to read files (including Utf8 errors) can be naturally
captured there as a new case.